refactor(uclient): fold chain_reader.go into event_processor and types
ChainReader moves next to its consumer (EventProcessor), the uint256
canonical encoder moves to types.go, and the near-no-op bytes32 encoder
is inlined at its single call site.
